Move user module model classes to the authentication module
