+++ modules/user/server-scala/src/main/scala/model/ConfirmRegistrationInput.scala
