+++ modules/user/server-scala/src/main/scala/model/ForgotPasswordInput.scala
