+++ modules/user/server-scala/src/main/scala/model/RegisterUserInput.scala
