+++ modules/user/server-scala/src/main/scala/model/auth/AuthInput.scala
@@ -1,4 +1,10 @@
-package model
+package model.auth
+
+import model.AuthCertificateInput
+import model.auth.facebook.AuthFacebookInput
+import model.auth.github.AuthGitHubInput
+import model.auth.google.AuthGoogleInput
+import model.auth.linkedin.AuthLinkedInInput
 
 case class AuthInput(
     certificate: Option[AuthCertificateInput],
