+++ modules/user/server-scala/src/main/scala/model/auth/AuthPayload.scala
@@ -1,7 +1,8 @@
-package model
+package model.auth
 
 import common.FieldError
 import jwt.model.Tokens
+import model.User
 
 case class AuthPayload(
     user: Option[User] = None,
