+++ modules/user/server-scala/src/main/scala/model/auth/facebook/AuthFacebookInput.scala
@@ -1,3 +1,3 @@
-package model
+package model.auth.facebook
 
 case class AuthFacebookInput(fbId: Option[String], displayName: Option[String])
