+++ modules/user/server-scala/src/main/scala/model/auth/github/AuthGitHubInput.scala
@@ -1,3 +1,3 @@
-package model
+package model.auth.github
 
 case class AuthGitHubInput(ghId: Option[String], displayName: Option[String])
