+++ modules/user/server-scala/src/main/scala/model/auth/google/AuthGoogleInput.scala
@@ -1,3 +1,3 @@
-package model
+package model.auth.google
 
 case class AuthGoogleInput(googleId: Option[String], displayName: Option[String])
