+++ modules/user/server-scala/src/main/scala/model/auth/linkedin/AuthLinkedInInput.scala
@@ -1,3 +1,3 @@
-package model
+package model.auth.linkedin
 
 case class AuthLinkedInInput(lnId: Option[String], displayName: Option[String])
